listitemmanager: Really remove section boundary on insert
Complete the fix introduced by b05000d8bdbe40771b59edd793574263ef6cb60f

It was made to handle the case of inserting at a tracked position which
lies immediately after the a section boundary, of this kind:

  ][

where ']' and '[' are matched FOOTER and HEADER, as in the testsuite.

But it doesn't handle the other 2 cases:

  ](
  )[

where ')' is UNMATCHED_FOOTER and '(' is UNMATCHED_HEADER. Indeed, I
got testsuite/gtk/listitemmanager to crash on these 2 scenarios.

Note: the ')(' case doesn't exist because, at rest, there are no
boundary tiles (and no header widget) between untracked sections.

So, generalize the previous fix and add comments to make it easier
to understand the logic.