dmabuf: Add Wayland information
Add device and scanout information for dmabuf formats.

While this is not a great API, there is no good alternative to
making this information available to applications that want to
negotiate dmabuf formats suitable for graphics offload.

The getters for Wayland-specific information are kept as
Wayland backend apis.

Tests included.