gsk/gpu: avoid Private usage for GskGpuImage
By using GskGpuImagePrivate for fields we require a runtime lookup of the
offset to add to the address for dereference. This appears to prevent LTO
from opperating successfully here.

Keep the API the same but also move these into static inline so that we
don't require extremely slow LTO builds to optimize further.