Add gsk_font_get_rendering
This function takes the font and scale, and determines what font
options to use, based on the passed down font rendering setting,
and the device pixel size of the font.

Currently, we use hinting if the font size is less than 31 pixels,
which corresponds to the size of an 11pt font at a dpi of 200.