Revert "gdk/frameclock: add mark when we discover frames may drop"
This reverts commit 84a304e66ee2f18ac1556350656e5e7135a6d810.

This produces marks that are confusing to me. They don't correlate
with actual gaps in the frame cycle and often overlap with regular
'window presented' marks. Also, the function we are emitting these
marks from is called from the get_frame_time getter, and we
definitely don't want to emit marks from there.