build: Handle in-tree copies of wayland protocols
Convert the array managing the Wayland protocols to contain
dictionary objects for each of the protocol definitions, in order
to make it easier to extend for extra keys in them.

And also add one such extra 'required' key, so that we can
optionally build protocols depending on the wayland-protocol
version detected at runtime. This is necessary to add fallbacks
for upstream protocols without necessarily bumping GTK dependency,
some versioning handling of private protocols was added too
for this purpose, so we can have in-tree copies of them.